Task 5 - Personal Finance Calculator

+++ python_introduction/finance_calculator.py
@@ -1,9 +1,9 @@
 
-monthly_income = int(input('Enter your monthly income: '))
-monthly_expense = int(input('Enter your total monthly expenses: '))
+monthly_income = float(input('Enter your monthly income: '))
+monthly_expense = float(input('Enter your total monthly expenses: '))
 
-monthly_saving = int(monthly_income - monthly_expense)
-projected_savings = int(monthly_saving * 12 + int(monthly_saving * 12 * 0.05))
+monthly_saving = monthly_income - monthly_expense
+projected_savings = monthly_saving * 12 + monthly_saving * 12 * 0.05
 
 print(f'Enter your monthly income: {monthly_income}')
 print(f'Enter your total monthly expenses: {monthly_expense}')
